+++ wikipron/extract/khm.py
@@ -3,10 +3,9 @@
 import itertools
 import typing
 
-import requests
+import requests_html
 
 from wikipron.extract.default import yield_pron
-
 
 if typing.TYPE_CHECKING:
     from wikipron.config import Config
@@ -17,7 +16,7 @@ _IPA_XPATH_SELECTOR = '//span[@class = "IPA" and @lang = "km"]'
 
 
 def extract_word_pron_khmer(
-    word: "Word", request: requests.Response, config: "Config"
+    word: "Word", request: requests_html, config: "Config"
 ) -> "Iterator[WordPronPair]":
     words = itertools.repeat(word)
     prons = yield_pron(request.html, _IPA_XPATH_SELECTOR, config)